refactor: move 'models' to 'types'
