+++ src/lib/types/configuration.ts
