+++ src/lib/types/image-table.ts
@@ -1,4 +1,4 @@
-import type { RepoImage } from '$lib/models/image.ts';
+import type { RepoImage } from '$lib/types/image';
 import type { ColumnDef } from '@tanstack/table-core';
 import { renderComponent } from '$lib/components/ui/data-table/index.js';
 import TagDropdownActions from '$lib/components/image-table/tag-dropdown-actions.svelte';
